Updates helper file to include new helpers
