+++ src/scripts/helpers/formControllerHelpers.ts
@@ -137,3 +137,90 @@ export async function handlePostEditForm(
     next(error);
   }
 }
+
+/**
+ * Handles validation errors and renders the form with error messages
+ * @param {Request} req - Express request object
+ * @param {Response} res - Express response object
+ * @param {string} caseReference - Case reference number
+ * @param {Record<string, unknown>} formFields - Form field values
+ * @returns {boolean} - Returns true if there were validation errors, false otherwise
+ */
+export function handleThirdPartyValidationErrors(
+  req: Request,
+  res: Response,
+  caseReference: string,
+  formFields: Record<string, unknown>
+): boolean {
+  const rawValidationResult = validationResult(req);
+
+  if (rawValidationResult.isEmpty()) {
+    return false;
+  }
+
+  const rawErrors = rawValidationResult.array();
+
+  const resultingErrors = rawErrors.map((error) => {
+    const fieldName = 'path' in error && typeof error.path === 'string' ? error.path : '';
+    const errorData = formatValidationError(error);
+
+    return {
+      fieldName,
+      inlineMessage: errorData.inlineMessage,
+      summaryMessage: errorData.summaryMessage,
+    };
+  });
+
+  const inputErrors = resultingErrors.reduce((errors: Record<string, string>, { fieldName, inlineMessage }) => {
+    if (inlineMessage.trim() !== '') {
+      errors[fieldName] = inlineMessage;
+    }
+    return errors;
+  }, {});
+
+  const errorSummaryList = resultingErrors.map(({ summaryMessage, fieldName }) => ({
+    text: summaryMessage,
+    href: `#${fieldName}`,
+  }));
+
+  const renderData = {
+    caseReference,
+    error: { inputErrors, errorSummaryList },
+    csrfToken: typeof req.csrfToken === 'function' ? req.csrfToken() : undefined,
+    currentThirdPartyFullName: formFields.thirdPartyFullName,
+    currentThirdPartyEmailAddress: formFields.thirdPartyEmailAddress,
+    currentThirdPartyContactNumber: formFields.thirdPartyContactNumber,
+    currentThirdPartySafeToCall: formFields.thirdPartySafeToCall,
+    currentThirdPartyAddress: formFields.thirdPartyAddress,
+    currentThirdPartyPostcode: formFields.thirdPartyPostcode,
+    currentThirdPartyRelationshipToClient: formFields.thirdPartyRelationshipToClient,
+    currentThirdPartyPassphraseSetUp: formFields.thirdPartyPassphraseSetUp,
+    currentThirdPartyPassphrase: formFields.thirdPartyPassphrase
+  };
+
+  res.status(BAD_REQUEST).render('case_details/third_party_details/add-client-third-party.njk', renderData);
+  return true;
+}
+
+/**
+ * Prepares third party data for API submission
+ * @param {Record<string, unknown>} formFields - Form field values
+ * @returns {object} - Formatted third party data for API
+ */
+export function prepareThirdPartyData(formFields: Record<string, unknown>): object {
+  return {
+    fullName: formFields.thirdPartyFullName,
+    emailAddress: formFields.thirdPartyEmailAddress,
+    contactNumber: formFields.thirdPartyContactNumber,
+    safeToCall: formFields.thirdPartySafeToCall !== '' ? formFields.thirdPartySafeToCall : true,
+    address: formFields.thirdPartyAddress,
+    postcode: formFields.thirdPartyPostcode,
+    relationshipToClient: {
+      selected: Array.isArray(formFields.thirdPartyRelationshipToClient) ? formFields.thirdPartyRelationshipToClient : [formFields.thirdPartyRelationshipToClient]
+    },
+    passphraseSetUp: {
+      selected: Array.isArray(formFields.thirdPartyPassphraseSetUp) ? formFields.thirdPartyPassphraseSetUp : [formFields.thirdPartyPassphraseSetUp],
+      passphrase: formFields.thirdPartyPassphrase
+    }
+  };
+}
